% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_Functions.R
\name{cdfw}
\alias{cdfw}
\title{Generate weights for bayesian distribution regression 3D plot}
\usage{
cdfw(mat, lgyt = 512, nrmc = NULL)
}
\arguments{
\item{mat}{matrix of draws from the bivariate distribution}

\item{lgyt}{the length of grid values of average fitted probabilities to use (default: 512)}

\item{nrmc}{the choice of normalising constant for the function \code{fnCDF} (default: NULL which corresponds to column sums of pdf weights, alternative: "MaxColSum" maximum column sum)}
}
\value{
val a matix with the same dimensions as mat for the the 3D plot
}
\description{
\code{cdfw} generates a matrix of weights required for dimension z of 3D plot
}
