% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_fringe.R
\docType{data}
\name{dat_fringe}
\alias{dat_fringe}
\title{Data - fringe}
\format{A data frame with 616 rows and 3 variables:
\describe{
\item{yy}{=1 if live in northeast}
\item{X2}{years work experience}
\item{X3}{squared of years work experience}
}}
\source{
\url{http://fmwww.bc.edu/ec-p/data/wooldridge/datasets.list.html}
}
\description{
A reduced version of the fringe data set on Jeffrey M. Wooldridge's Econometrics texts' website
}
\keyword{data}
