% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression Functions.R
\name{distreg}
\alias{distreg}
\title{Bayesian distribution regression}
\usage{
distreg(thresh, data0, MH = "IndepMH", ...)
}
\arguments{
\item{thresh}{threshold value that is used to binarise the continuous dependent variable}

\item{data0}{original data set with the first column being the continuous dependent variable}

\item{MH}{metropolis-hastings algorithm to use; default:"IndepMH", alternative "RWMH"}

\item{...}{any additional inputs to pass to the MH algorithm}
}
\value{
fitob a vector of fitted values corresponding to the distribution at threshold thresh
}
\description{
\code{distreg} computes the distribution at a threshold value
}
\examples{
data0=DAT_ALL; qnts<-quantile(data0[,1],c(0.5,0.25,0.5,0.75,0.95));
distob<- distreg(qnts[1],data0)

}
