% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression_Functions.R
\name{distreg_cfa}
\alias{distreg_cfa}
\title{Counterfactual bayesian distribution regression}
\usage{
distreg_cfa(thresh, data0, MH = "IndepMH", cft, cfIND, ...)
}
\arguments{
\item{thresh}{threshold value that is used to binarise the continuous dependent variable}

\item{data0}{original data set with the first column being the continuous dependent variable}

\item{MH}{metropolis-hastings algorithm to use; default:"IndepMH", alternative "RWMH"}

\item{cft}{counterfactual treatment; this can be a vector or matrix of counterfactual covariates}

\item{cfIND}{the column index(indices) of covariates to replace with \code{cft} in \code{data0}}

\item{...}{any additional inputs to pass to the MH algorithm}
}
\value{
robj a list of a vector of fitted values corresponding to the distribution at threshold thresh
}
\description{
\code{distreg} computes the distribution and its counterfactual at a threshold value
}
\examples{
# data0=DAT_ALL; qnt<-quantile(data0[,1],0.5) #at the median of y
# cfIND=c(2,3) #Note: the first column is the dependent variable. 
# cft=0.95*data0[,cfIND] # a decrease by 5\%
# dist_cfa<- distreg_cfa(qnt,data0,cft,cfIND,iter=3000)

}
