% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis Functions.R
\name{EsF}
\alias{EsF}
\title{Get distribution of expected shortfall}
\usage{
EsF(full_out, y_0, tau = 0.05)
}
\arguments{
\item{full_out}{full matrix of output. The rows correspond to threshold
values y_0 and the columns correspond to the number of MCMC draws}

\item{y_0}{vector of threshold \code{y} values. It must have length equal to 
the number of rows of \code{full_out}
\code{"mode"}, or \code{"quantile"}. The default is \code{NULL} which returns the median}

\item{tau}{the probability tau=0.05 is the probability of VaR}
}
\value{
val this is the distribution of the VaR
}
\description{
This function enables one to generate the distribution of the VaR
}
