% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{fitdist}
\alias{fitdist}
\title{The distribution of mean fitted logit probabilities}
\usage{
fitdist(Matparam, data)
}
\arguments{
\item{Matparam}{an M x k matrix of parameter draws, each being 1 x k}

\item{data}{dataframe used to obtain Matparam}
}
\value{
dist fitted distribution
}
\description{
\code{fitdist} function generates a vector of mean fitted probabilities that constitute the 
distribution
}
