% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{fitlogit}
\alias{fitlogit}
\title{Fitted logit probabilities}
\usage{
fitlogit(pars, data)
}
\arguments{
\item{pars}{vector of parameters}

\item{data}{data frame. The first column of the data frame ought to be the binary dependent 
variable}
}
\value{
vec vector of fitted logit probabilities
}
\description{
\code{fitlogit} obtains a vector of fitted logit probabilities given parameters (pars)
and data
}
