% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis Functions.R
\name{fnCDF}
\alias{fnCDF}
\title{Compute bivariate distribution}
\usage{
fnCDF(vec, nrmc = NULL)
}
\arguments{
\item{vec}{vector of input}

\item{nrmc}{normalising constant}
}
\value{
val a vector of normalised cumulative weights
}
\description{
\code{fnCDF} generates the cumulative weights of a vector given a normalising constant
}
\examples{
kd<- matrix(runif(12,0,10),ncol = 3); apply(kd,2,fnCDF,nrmv=max(apply(kd,2,sum)))
}
