% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis Functions.R
\name{getmode}
\alias{getmode}
\title{Get mode of a vector}
\usage{
getmode(vec)
}
\arguments{
\item{vec}{vector of values from the continuous distribution}
}
\value{
mval the mode of the continuous distribution
}
\description{
\code{getmode} is used to obtain the mode from a vector representing a constinuous distribution.
For a discrete variable, use the function \code{table()}
}
