% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{getsmooth}
\alias{getsmooth}
\title{Smooth marginal conditional distribution}
\usage{
getsmooth(vec, Step = NULL)
}
\arguments{
\item{vec}{a vector of representing the marginal conditional quantile}

\item{Step}{length of moving window used to calculate the moving mean}
}
\value{
val vector of smoothed marginal conditional quantile
}
\description{
\code{getsmooth} uses moving means to minimise kinks in marginal conditional distributions
}
