% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{getsmooth.spline}
\alias{getsmooth.spline}
\title{Smooth marginal conditional distribution using smooth.spline()}
\usage{
getsmooth.spline(x, y, CI = NULL, cv = TRUE, ...)
}
\arguments{
\item{x}{a vector of representing the marginal conditional quantile}

\item{y}{length of moving window used to calculate the moving mean}

\item{CI}{(optional) a list of \code{CI$lb} and \code{CI$up} of lower and upper confidence
intervals of \code{y}}

\item{cv}{logical for cross-validation to be passed to \code{smooth.spline()}}

\item{...}{additional paramters to be passed to \code{smooth.spline()}}
}
\value{
out a list with the following components:

x output vector x

y output vector y

lb lower confidence band if \code{CI} is provided

ub upper confidence band if \code{CI} is provided
}
\description{
\code{getsmooth.spline} uses the function \code{smooth.spline()} in the 
\code{stats} package for smoothing
}
