% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{indicat}
\alias{indicat}
\title{Indicator function}
\usage{
indicat(y, y0)
}
\arguments{
\item{y}{vector y}

\item{y0}{threshold value y0}
}
\value{
val
}
\description{
This function creates 0-1 indicators for a given threshold y0 and vector y
}
