% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{lapl_aprx2}
\alias{lapl_aprx2}
\title{An \code{optim}-based function maximiser}
\usage{
lapl_aprx2(start, func, ...)
}
\arguments{
\item{start}{the starting parameter values}

\item{func}{the function to be maximised}

\item{...}{other arguments to be passed to func}
}
\value{
val list of mode and variance-covariance matrix
}
\description{
\code{lapl_aprx2} is used to take a laplace approximation of the posterior distribution
}
