% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{LogitLink}
\alias{LogitLink}
\title{Logit function}
\usage{
LogitLink(x)
}
\arguments{
\item{x}{Random variable}
}
\value{
val Probability value from the logistic function
}
\description{
This is the link function for logit regression
}
