% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_Functions.R
\name{mcondist}
\alias{mcondist}
\title{Get marginal conditional distribution}
\usage{
mcondist(full_out, diststat = NULL, tau = 0.5)
}
\arguments{
\item{full_out}{full matrix of output. The rows correspond to threshold
values y_0 and the columns correspond to the number of MCMC draws}

\item{diststat}{distributional statistic; one can choose \code{"mean"}, \code{"median"}
\code{"mode"}, or \code{"quantile"}. The default is \code{NULL} which returns the median}

\item{tau}{the quantile. tau=0.5 corresponds to the median. this option is turned off
when \code{diststat} is set to another distributional statistic.}
}
\value{
val this is a vector of the marginal conditional distribution
}
\description{
This function enables one to obtain the marginal conditional distribution at a
chosen distributional statistic, eg. mean, mode, median, or any quantile
}
