% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression_Functions.R
\name{par_distreg2}
\alias{par_distreg2}
\title{Parallel compute bayesian distribution regression}
\usage{
par_distreg2(thresh, data0, fn = distreg, ...)
}
\arguments{
\item{thresh}{vector of threshold values. ensure the min and the max generate enough zeros and ones.}

\item{data0}{the original data set with a continous dependent variable in the first column}

\item{fn}{bayesian distribution regression function. the default is distreg provided in the package}

\item{...}{any additional input parameters to pass to fn}
}
\value{
mat a G x M matrix of output (G is the length of thresh, M is the number of draws)
}
\description{
This version of the function uses the foreach package
}
\examples{
# data0=DAT_ALL; qnts<-quantile(data0[,1],c(0.5,0.25,0.5,0.75,0.95))
# out2<- par_distreg2(qnts,data0)

}
