% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{prior_n}
\alias{prior_n}
\title{Normal Prior distribution}
\usage{
prior_n(pars, mu, sig, Log = FALSE)
}
\arguments{
\item{pars}{parameter values}

\item{mu}{mean value of each parameter value}

\item{sig}{standard deviation of each parameter value}

\item{Log}{logical to take the log of prior or not (defaults to FALSE)}
}
\value{
val Product of probability values for each parameter
}
\description{
This normal prior distribution is a product of univariate N(mu,sig)
}
\examples{
prior_n(rep(0,6),0,10,Log = TRUE) #log of prior
prior_n(rep(0,6),0,10,Log = FALSE) #no log

}
