% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{prior_u}
\alias{prior_u}
\title{Uniform Prior distribution}
\usage{
prior_u(pars)
}
\arguments{
\item{pars}{parameter values}
}
\value{
val value of joint prior =1 for the uniform prior
}
\description{
This uniform prior distribution proportional to 1
}
