% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{psimval}
\alias{psimval}
\title{A Bayesian "p-value"}
\usage{
psimval(
  DF,
  DFmat,
  ialpha = 0.05,
  eps = 1e-04,
  typeband = c("sym", "asym", "OM")
)
}
\arguments{
\item{DF}{the target vector needed to construct the confidence band}

\item{DFmat}{the matrix of draws of the distribution, rows correspond to 
elements in \code{DF}}

\item{ialpha}{a starting value of alpha to be searched on the interval (0,1)}

\item{eps}{steps of the grid on the interval (0,1) to search}

\item{typeband}{the type of confidence band to use: "sym" for \code{\link{simcnfB}}, 
"asym" for \code{\link{asymcnfB}}, or "OM" for \code{\link{jntCBOM}} are supported.}
}
\value{
alf the "p-value"
}
\description{
\code{psimval} finds a minimum alpha in the interval (0,1) such that the 1-alpha simultaneous 
confidence band excludes the vector of DF. This is intended to be a summary value for a joint
hypothesis test of equality.
}
\examples{
set.seed(14); m=matrix(rbeta(500,1,4),nrow = 5) + 1:5
(pval<- round(psimval(DF=rep(0.1,5),DFmat = m,ialpha=0.05,typeband="OM"),3))

}
