% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{asymcnfB}
\alias{asymcnfB}
\title{Asymmetric simultaneous bayesian confidence bands}
\usage{
asymcnfB(DF, DFmat, alpha = 0.05, scale = FALSE)
}
\arguments{
\item{DF}{the target distribution/quantile function as a vector}

\item{DFmat}{the matrix of draws of the distribution, rows correspond to 
indices elements in \code{DF}}

\item{alpha}{level such that \code{1-alpha} is the desired probability of coverage}

\item{scale}{logical for scaling using the inter-quartile range}
}
\value{
cstar - a constant to add and subtract from DF to create 
 confidence bands if no scaling=FALSE else a vector of length DF.
}
\description{
\code{asymcnfB} obtains asymmetric bayesian distribution confidence bands
}
