% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{lapl_aprx}
\alias{lapl_aprx}
\title{Laplace approximation of posterior to normal}
\usage{
lapl_aprx(y, x)
}
\arguments{
\item{y}{the binary dependent variable y}

\item{x}{the matrix of independent variables.}
}
\value{
val A list of mode variance-covariance matrix, and scale factor for
proposal draws from the multivariate normal distribution.
}
\description{
This function generates mode and variance-covariance for a normal proposal
distribution for the bayesian logit.
}
\examples{
dat=genmle::dat_mroz ;y=dat$y; x = dat[,-1]; gg<- lapl_aprx(y,x)

}
