% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression Functions.R
\name{par_distreg}
\alias{par_distreg}
\title{Parallel compute bayesian distribution regression}
\usage{
par_distreg(thresh, data0, fn = distreg, ...)
}
\arguments{
\item{thresh}{vector of threshold values. ensure the min and the max generate enough zeros and ones.}

\item{data0}{the original data set with a continous dependent variable in the first column}

\item{fn}{bayesian distribution regression function. the default is distreg provided in the package}

\item{...}{any additional input parameters to pass to fn}
}
\value{
mat a G x M matrix of output (G is the length of thresh, M is the number of draws)
}
\description{
This function uses parallel computation to compute bayesian distribution regression for a given
vector of threshold values and a data (with first column being the continuous dependent variable)
}
\examples{
data0=DAT_ALL; qnts<-quantile(data0[,1],c(0.5,0.25,0.5,0.75,0.95))
out<- par_distreg(qnts,data0)

}
