% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{parLply}
\alias{parLply}
\title{Parallel compute}
\usage{
parLply(vec, fn, type = "FORK", no_cores = 1, ...)
}
\arguments{
\item{vec}{vector of inputs over which to parallel compute}

\item{fn}{the function}

\item{type}{this option is set to "FORK", use "PSOCK" on windows}

\item{no_cores}{the number of cores to use. Defaults at 1}

\item{...}{extra inputs to \code{fn()}}
}
\value{
out parallel computed output
}
\description{
\code{parLply} uses \code{parlapply} from the \code{parallel} package with 
a function as input
}
