% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{parLply}
\alias{parLply}
\title{Parlapply a function}
\usage{
parLply(vec, fn, type = "FORK", ...)
}
\arguments{
\item{vec}{vector of inputs over which to parallel compute}

\item{fn}{the function}

\item{type}{this option is set to "FORK", use "PSOCK" for windows}

\item{...}{extra inputs to \code{fn}}
}
\value{
out parallel computed output
}
\description{
\code{paRLply} parlapply from the \code{parallel} package with a function as input
}
