% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{quant_bdr}
\alias{quant_bdr}
\title{Quantile conversion of a bayesian distribution matrix}
\usage{
quant_bdr(taus, thresh, mat)
}
\arguments{
\item{taus}{a vector of quantile indices}

\item{thresh}{a vector of threshold values used in a \code{par_distreg()} type function}

\item{mat}{bayesian distribution regression output matrix}
}
\value{
qmat matrix of quantile distribution
}
\description{
\code{quant_bdr} converts a bayesian distribution regression matrix from \code{par_distreg()}
output to a matrix of quantile distribution.
}
