% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metropolis_Hastings_Algorithms.R
\name{RWMH}
\alias{RWMH}
\title{Random Walk Metropolis-Hastings Algorithm}
\usage{
RWMH(data, propob = NULL, posterior = NULL, iter = 15000, burn = 1000,
  start = NULL, prior = "Normal", mu = 0, sig = 10)
}
\arguments{
\item{data}{data required for the posterior distribution}

\item{propob}{a list of mean and variance-covariance of the normal proposal distribution
(default: NULL i.e. internally generated)}

\item{posterior}{the posterior distribution. It is set to null in order to use the logit posterior.
The user can specify log posterior as a function of parameters and data (pars,data)}

\item{iter}{number of random draws desired (default: 15000)}

\item{burn}{burn-in period for the Random Walk MH algorithm (default: 1000)}

\item{start}{starting values of parameters for the MH algorithm.
It is automatically generated but the user can also specify.}

\item{prior}{the prior distribution (default: "Normal", alternative: "Uniform")}

\item{mu}{the mean of the normal prior distribution (default:0)}

\item{sig}{the variance of the normal prior distribution (default:10)}
}
\value{
val a list of matrix of draws pardraws and the acceptance rate
}
\description{
\code{RWMH} computes random draws of parameters using a specified proposal distribution.
The default is the normal distribution
}
\examples{
data = genmle::dat_mroz
propob<- lapl_aprx(data[,1],data[,-1])
RWMHob<- RWMH(data=data,propob) # prior="Normal"
system.time(RWMHob<- RWMH(data=data,propob)) # time operation
RWMHob<- RWMH(data=data,prior="Uniform")
RWMHob<- RWMH(data=data,iter=50000) # a higher number of draws

}
