% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{simcnfB}
\alias{simcnfB}
\title{Simultaneous bayesian confidence bands}
\usage{
simcnfB(DF, DFmat, alpha = 0.05)
}
\arguments{
\item{DF}{the target distribution as a vector}

\item{DFmat}{the matrix of draws of the distribution, rows correspond to 
indices elements in \code{DF}}

\item{alpha}{level such that \code{1-alpha} is the desired probability of coverage}
}
\value{
cstar - a constant to add and subtract from DF to create 
the confidence bands.

uCB - upper band
}
\description{
\code{simcnfB} obtains bayesian distribution confidence bands
}
